Switch special page to be a QueryPage rather than using an IndexPager
Because the page_assessments table has a multi-column PK, IndexPager
doesn't have a field to sort by. So we switch to making the special
page a subclass of QueryPage instead. This makes pagination much easier
but with the downside of there being no pre-built sortable column
headers. So, we borrow the TablePager CSS for this.

The search query parameters remain the same ('project', 'page_title',
'namespace', and 'sort'), but the sort direction changes to 'dir'.

Bug: T152756
Change-Id: I6b70f42a30b4e1a0a4c6ff86986c835b96afeecd