Fixing special page time-out
Previously, the query was generating a list several million items
long, as it was adding a project name for every assessment in the
page_assessments table. Now that we have a maintanence script that
removes unused projects every 24 hours, there really isn't any
reason to pro-actively filter for only used projects, so we might
as well simplify the query (especially since the results are only
used for auto-suggestion anyway).

Bug: T156198
Change-Id: I28c3bb5b64ff4a91a3dd998daea7857959c780a3