Protect against empty array causing sql syntax error
In some environments, the array of projects that is built by the
script can be empty. When that happens, the resulting sql query
ends up having a syntax error in its NOT IN clause.

This change checks to ensure that the array of projects is not
empty before doing the subsequent query. If it is empty, it
outputs and then returns to stop execution.

Bug: T219935
Change-Id: Iede2021e99fba778f1ab03593b0fbdf06b66661b