Add CirrusSearch integration to enable searching by WikiProject
Use WeightedTagsUpdater to load project associations to ElasticSearch
after they are updated in the db during LinksUpdate. The importance
parameter is used as the weight for ranking search results.

An "inproject:" search keyword is added to filter results to pages
associated with the specified project(s).

Results for inproject searches may be incomplete until a full reindexing
has occurred.

Bug: T378868
Change-Id: I143a6c3e9e4cff075e97474bafa04072d60e80bd