Do not update the search index if the assessment did not change
Followup on I53df3cda65e680763ef7ce36b55fa6eb649e6e30 which assumed that
all assessments in $toUpdate required a change. This might not be true
if all its metadata remain unchanged.
Change the semantic of PageAssessmentsDAO::updateRecord to inform the
called whether an update was performed or not.

Switch the semantic of "$nochange" to "$changed" to avoid the double
negation "if ( !$nochange )".

Change-Id: Iff1c63b2a013aa8b78c6b018593c1f48ca4845ef